// components/property-form/types.ts
export type ListingDraft = {
  title: string;
  location: string;
  price: string;
  configuration: string;
  area: string;
  availability: string;
  notes: string;
  photoUrls: string[];
  intent: string;
  category: string;
  propertyType: string;
  officeType: string;
  retailType: string;
  shopLocation: string;
  shopWashroom: string;
  parkingType: string;
  entranceWidth: string;
  ceilingHeight: string;
  bookingAmount: string;
  industryType: string;
  hospitalityType: string;
  plotType: string;
  storageType: string;
  minWorkstations: string;
  maxWorkstations: string;
  cabins: string;
  meetingRooms: string;
  washroomAvailability: string;
  conferenceRoom: string;
  receptionArea: string;
  pantryType: string;
  facilityParking: string;
  centralAc: string;
  lifts: string;
  parkingAvailability: string;
  preLeased: string;
  currentRent: string;
  leaseTenure: string;
  annualRentIncrement: string;
  leasedTo: string;
  dgUpsIncluded: boolean;
  city: string;
  locality: string;
  subLocality: string;
  society: string;
  houseNumber: string;
  bedrooms: string;
  bathrooms: string;
  balconies: string;
  additionalRooms: string[];
  furnishing: string;
  totalFloors: string;
  propertyFloor: string;
  possession: string;
  propertyAge: string;
  possessionBy: string;
  areaType: string;
  carpetArea: string;
  plotArea: string;
  builtUpArea: string;
  superBuiltUpArea: string;
  facing: string;
  powerBackup: string;
  flooring: string;
  coveredParking: string;
  openParking: string;
  waterSource: string;
  ownership: string;
  postedAs: string;
  negotiable: boolean;
  allInclusive: boolean;
  priceDetails: string;
  maintenance: string;
  deposit: string;
  preferredTenant: string;
  preferredTenants: string[];
  availableFrom: string;
  brokerContact: string;
  chargesExcluded: boolean;
  plotLength: string;
  plotBreadth: string;
  boundaryWall: string;
  openSides: string;
  floorsAllowed: string;
  constructionDone: string;
  constructionTypes: string[];
  approvedBy: string;
  amenities: string[];
};

export type ListingChange = <K extends keyof ListingDraft>(
  field: K,
  value: ListingDraft[K],
) => void;